requirements.txt
Flask==3.1.0
flask_cors==5.0.1
mediapipe==0.10.14
numpy==2.2.4
opencv_contrib_python==4.10.0.84
opencv_python==4.11.0.86
Pillow==11.2.1
tensorflow==2.17.0
tensorflow_intel==2.17.0
